Get OTB function argument list

retrieve the choosen function and returns a full argument list with the default settings

parseOTBFunction(algo = NULL, gili = NULL)

Arguments

algo
either the number or the plain name of the `OTB` algorithm that is wanted. Note the correct (of current/choosen version) information is probided by `parseOTBAlgorithms()`

gili
optional list of avalailable `OTB` binaries if not provided `linkOTB()` is called

Examples

if (FALSE) {
otbLink<-link2GI::linkOTB()
if (otbLink$exist) {

## parse all modules
algos<-parseOTBAlgorithms(gili = otbLink)

## take edge detection
cmdList<-parseOTBFunction(algo = algos[27],gili = otbLink)
## print the current command
print(cmdList)
}
}
##+##