Bridges to the GI-World

A straightforward helper tool for linking GI/RS functionality to R. The goal of the package is to correctly initialize both the existing wrapper packages rgrass7 and RSAGA and to smoothly enable the necessary system variables and path parameters for a direct access of the binaries via direct system calls on all operating systems. In particular, rgrass7 and RSAGA can cause severe problems during initialization of parallel installations of GRASS GIS or SAGA GIS under the Windows operating system(s). link2GI tries to set the correct system settings and returns if system calls are required the necessary paths and command strings.
Furthermore the package provides a linkage to the Orfeo Toolbox (OTB) software. Due to the difficulties linking the correct GDAL binaries a new system wide search for GDAL binaries is implemented.

Finally there are some usefull functions for creating project folder structures and project environments. To deal with the not always consistent API-calls of OTB a list based command parser and generator is provided.

Details

Functions for linking GI/RS functionality to R

Note

To utilize the power of the open source GI tools from within R you need to installthem first. The link2GI package just tries to generate correct environment settings as system and path variables for the most of the known issues. As a first promising opportunity to do fullfil most of the requirements you may install QGIS, GRASS GIS 7.x and SAGA-GIS following the excellent installation instructions of the RQGIS package will have a good first try to ensure a smooth working environment.

For a broad number of problems you may just use the RQGIS package as wrapper for the functionalitythat can be reached via QGIS. link2GI is tested under Windows 7/10 as well as on the Ubuntu/Debian Linux distributions. The OSX operation system should run but is not tested.